add ui and polling

+++ web/src/main/java/com/taobao/tae/ewall/dashboard/pipeline/PipelineController.java
@@ -1,8 +1,12 @@
 package com.taobao.tae.ewall.dashboard.pipeline;
 
 
+import com.taobao.tae.ewall.dashboard.project.ProjectUtil;
+import com.taobao.tae.ewall.project.ProjectGridImpl;
+import com.taobao.tae.ewall.project.ProjectDO;
 import com.taobao.tae.ewall.pipeline.PipelineDO;
 import com.taobao.tae.ewall.pipeline.PipelineGroupDO;
+import com.taobao.tae.ewall.service.ProjectService;
 import com.taobao.tae.ewall.service.PipelineGroupService;
 import com.taobao.tae.ewall.service.PipelineService;
 import org.apache.commons.lang.StringUtils;
@@ -25,6 +29,8 @@ import java.util.List;
 public class PipelineController {
 
     @Resource
+    ProjectService projectService;
+    @Resource
     PipelineService pipelineService;
     @Resource
     PipelineGroupService pipelineGroupService;
@@ -38,36 +44,42 @@ public class PipelineController {
     @RequestMapping("pipelinelist")
     public String getAllPipelinesSortByGroup(ModelMap model) {
         HashMap<String, List<PipelineDO>> pipelinesMap = new HashMap<String, List<PipelineDO>>();
+        List<ProjectDO> projectList = projectService.getAllProjects();
         List<PipelineGroupDO> groupDOs = pipelineGroupService.getAllGroups();
         for (int i = 0; i < groupDOs.size(); i++) {
             List<PipelineDO> pipelineDOs = pipelineService.findPipelinesByGroupId(groupDOs.get(i).getId());
             pipelinesMap.put(groupDOs.get(i).getName(), pipelineDOs);
         }
         model.addAttribute("pipelinesMap", pipelinesMap);
+        model.addAttribute("projectList", projectList);
         return "dashboard/pipeline/pipelinelist";
     }
 
     @RequestMapping(value = "pipelineadd", method = RequestMethod.POST)
-    public String addGroup(String name, String groupName, String description, ModelMap model) {
+    public String addPipeline(String name, String groupName, Integer headProjectId, String description, String display, ModelMap model) {
         PipelineGroupDO pipelineGroupDO = pipelineGroupService.findByName(groupName);
         if (StringUtils.isNotBlank(name)) {
             PipelineDO pipelineDO = new PipelineDO();
             pipelineDO.setName(name);
+            pipelineDO.setHeadProjectId(headProjectId);
             pipelineDO.setDescription(description);
+            pipelineDO.setDisplay(Boolean.valueOf(display));
             pipelineDO.setGroupId(pipelineGroupDO.getId());
-            pipelineService.addPipeline(pipelineDO);
+            Integer pipelineId = pipelineService.addPipeline(pipelineDO);
+            pipelineService.initPipeline(pipelineId);
         }
         return "redirect:pipelinelist";
     }
 
     @RequestMapping(value = "pipelineedit", method = RequestMethod.POST)
-    public String editGroup(Integer id, String name,Integer groupId, String description, ModelMap model) {
+    public String editPipeline(Integer id, String name, Integer groupId, String description, String display, ModelMap model) {
         if (id != null && StringUtils.isNotBlank(name)) {
             PipelineDO pipelineDO = new PipelineDO();
             pipelineDO.setId(id);
             pipelineDO.setName(name);
             pipelineDO.setGroupId(groupId);
             pipelineDO.setDescription(description);
+            pipelineDO.setDisplay(Boolean.valueOf(display));
             Boolean result = pipelineService.updatePipeline(pipelineDO);
             model.addAttribute("editPipelineResult", result);
         }
@@ -76,6 +88,9 @@ public class PipelineController {
 
     @RequestMapping(value = "pipelinedelete", method = RequestMethod.GET)
     public String deletePipeline(Integer id, ModelMap model) {
+        if (id == null) {
+            return "redirect:pipelinelist";
+        }
         Boolean result = pipelineService.deleteById(id);
         model.addAttribute("deletePipelineResult", result);
         if (result) {
@@ -85,4 +100,32 @@ public class PipelineController {
         }
     }
 
+    /**
+     * 初始化 pipeline 管道中的 project 关系
+     *
+     * @param pipelineId
+     * @param model
+     * @return
+     */
+    @RequestMapping(value = "pipelinesetting", method = RequestMethod.GET)
+    public String pipelineSetting(Integer pipelineId, ModelMap model) {
+        if (pipelineId == null) {
+            return "redirect:pipelinelist";
+        }
+        PipelineDO pipelineDO = pipelineService.findPipelineById(pipelineId);
+        ProjectDO startProject = projectService.getProjectById(pipelineDO.getHeadProjectId());
+        List<ProjectDO> allProject = projectService.getAllProjects();
+        ProjectGridImpl projectGrid = new ProjectGridImpl(startProject, allProject);
+        ProjectUtil projectUtil = new ProjectUtil();
+
+        model.addAttribute("projectGrid", projectGrid);
+        model.addAttribute("columnSize", projectGrid.getColumns());
+        model.addAttribute("rowSize", projectGrid.getRows());
+        model.addAttribute("allProject", allProject);
+        model.addAttribute("pipelineId", pipelineId);
+        model.addAttribute("projectService", projectService);
+        model.addAttribute("projectUtil", projectUtil);
+
+        return "dashboard/pipeline/pipelinesetting";
+    }
 }
